requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async
aiohttp>=3.9.0
websockets>=12.0

# Data
pandas>=2.1.0
numpy>=1.26.0
pandas-ta>=0.3.14b0
redis>=5.0.0
yfinance>=0.2.36
fredapi>=0.5.2

# Trading
python-binance>=1.0.19
ccxt>=4.2.0

# LLM
anthropic>=0.18.0

# Utils
structlog>=24.1.0
tenacity>=8.2.0
cachetools>=5.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Dev
ruff>=0.3.0
mypy>=1.8.0
bandit>=1.7.0
